Changed to have same endpoint
