+++ api/v1/requests/index.js
@@ -8,7 +8,7 @@ module.exports = {
   EventCreationRequest: require('./EventCreationRequest'),
   LocationCreationRequest: require('./LocationCreationRequest'),
   MentorRequest: require('./MentorRequest'),
-  RecruiterInterestUpdateRequest: require('./RecruiterInterestUpdateRequest'),
+  UpdateRecruiterInterestRequest: require('./UpdateRecruiterInterestRequest'),
   RecruiterInterestRequest: require('./RecruiterInterestRequest'),
   ProjectRequest: require('./ProjectRequest'),
   ProjectMentorRequest: require('./ProjectMentorRequest'),
